requirements.txt
pytesseract
opencv-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
